Export all types from an index file
