+++ toolkit/src/components/cut/copy-block/types.ts
@@ -1,0 +1,7 @@
+export interface CopyBlockSignature {
+  Args: {
+    clipboardText: string;
+    timeout?: number;
+    onSuccess(): void;
+  };
+}
