+++ toolkit/src/components/cut/filter-bar/filter/index.ts
@@ -3,28 +3,7 @@
  */
 
 import Component from '@glimmer/component';
-import { FilterConfig, ToggleArgs } from '..';
-
-interface FilterInterface {
-  Args: {
-    name: string;
-    config: FilterConfig;
-    localConfig: FilterConfig;
-    toggle: (toggle: ToggleArgs) => void;
-    softToggle: (toggle: ToggleArgs) => void;
-    applyFilter: (name: string) => void;
-    clearPendingFilter: () => void;
-    isChecked: (
-      localConfig: FilterConfig,
-      name: string,
-      value: unknown
-    ) => boolean;
-    isMultiSelect?: boolean;
-    batch?: boolean;
-    dropdown?: unknown;
-    listPosition?: 'bottom-left' | 'bottom-right' | 'top-left' | 'top-right';
-  };
-}
+import { FilterSignature, ToggleArgs } from 'src/types';
 
 /**
  * `Cut::FilterBar::Filter` is a wrapper of the HDS::Dropdown component that exposes wrapped dropdown list items
@@ -43,7 +22,7 @@ interface FilterInterface {
  * @class Cut::FilterComponent
  *
  */
-export default class FilterComponent extends Component<FilterInterface> {
+export default class FilterComponent extends Component<FilterSignature> {
   get toggle(): (toggle: ToggleArgs) => void {
     return this.args.batch ? this.args.softToggle : this.args.toggle;
   }
