+++ toolkit/src/components/cut/list-item/service-instance/index.ts
@@ -3,44 +3,9 @@
  */
 
 import Component from '@glimmer/component';
-import {
-  ExternalSource,
-  HealthCheck,
-} from '../../../../utils/service-list-item';
+import { ServiceInstanceListItemSignature } from '../types';
 
-export interface CutServiceInstance {
-  name?: string;
-  metadata: {
-    healthCheck: {
-      node?: HealthCheck;
-      service?: HealthCheck;
-    };
-    tags: string[];
-    servicePortAddress?: string;
-    serviceSocketPath?: string;
-    node?: string;
-    externalSource?: ExternalSource;
-    connectedWithProxy?: boolean;
-  };
-}
-
-interface ComponentSignature {
-  Args: {
-    // List item args;
-    href?: string;
-    isHrefExternal?: boolean;
-    route?: string;
-    isRouteExternal?: boolean;
-    query?: object;
-    replace?: string;
-    onClick(): void;
-
-    // Service args
-    service: CutServiceInstance;
-  };
-}
-
-export default class ServiceInstanceListItemComponent extends Component<ComponentSignature> {
+export default class ServiceInstanceListItemComponent extends Component<ServiceInstanceListItemSignature> {
   get isAllHealthy() {
     const { healthCheck } = this.args.service.metadata;
     const serviceHealthy = healthCheck.service
