+++ toolkit/src/components/cut/list/types.ts
@@ -1,0 +1,13 @@
+export interface PaginationSignature {
+  Args: {
+    nextCursor?: string;
+    prevCursor?: string;
+    pageSizes?: number[];
+    currentPageSize?: number;
+    queryFunction?: (page: string) => {
+      [key: string]: string | number | unknown;
+    };
+    onPageChange?: (page: string) => void;
+    onPageSizeChange?: (size: number) => void;
+  };
+}
