+++ toolkit/src/components/cut/metadata/types.ts
@@ -1,0 +1,16 @@
+import { ExternalSource } from 'src/types';
+
+export interface MetadataHealthCheckBadgeSetSignature {
+  Args: {
+    type: string;
+    successCount?: number;
+    criticalCount?: number;
+    warningCount?: number;
+  };
+}
+
+export interface MetadataExternalSourceSignature {
+  Args: {
+    externalSource: ExternalSource;
+  };
+}
