+++ toolkit/src/components/cut/text-with-icon/types.ts
@@ -1,0 +1,9 @@
+export interface TextWithIconSignature {
+  Element: HTMLDivElement;
+  Args: {
+    icon: string;
+    text: string;
+    connection?: string | undefined;
+    iconColor?: string | undefined;
+  };
+}
